# meme_backend/src/main/resources/openapi/api.yml
openapi: 3.0.1
info:
  title: Meme backend
  version: 0.0.1
...
servers:
  - url: http://localhost:8080/api
    description: Generated server url
paths:
  /v1/postobject/{id}:
    get:
      summary: Returns post object
      tags:
        - "PostObject"
      operationId: getPostObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Post id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostObjectFullGet'
    put:
      summary: Updates one of the files/description in post
      security:
        - BearerAuth:
            - USER_EDIT
            - MODERATOR_EDIT
      tags:
        - "PostObject"
      operationId: updatePostObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of post object
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
                - description
              type: object
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
                  description: Description of file in post
                  example: Description of file in post
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostObjectFullGet'
    delete:
      summary: Deletes post object by id
      security:
        - BearerAuth:
            - USER_DELETE
            - MODERATOR_DELETE
      tags:
        - "PostObject"
      operationId: deletePostObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of post object
      responses:
        '200':
          description: OK
...
  /v1/tag/{id}/{tag}:
    put:
      summary: Updates tag
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "Tag"
      operationId: changeTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Tag id
        - name: tag
          in: path
          required: true
          schema:
            type: string
            example: Cat
            description: Tag value
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TagGet'
  /v1/profilepicture/{id}:
    get:
      summary: Returns profile picture by given id
      tags:
        - "ProfilePhoto"
      operationId: getProfilePicture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProfilePictureGet'
    put:
      summary: Updates profile picture
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "ProfilePhoto"
      operationId: changeDefaultProfilePicture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Profile picture id
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProfilePictureGet'
    delete:
      summary: Removes profile picture
      security:
        - BearerAuth:
            - MODERATOR_EDIT
      tags:
        - "ProfilePhoto"
      operationId: removeProfilePicture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Profile picture id
      responses:
        '200':
          description: OK
...
  /v1/poststatistic/{postId}/vote/{vote}:
    put:
      summary: Gives vote for post could be positive or negative
      security:
        - BearerAuth:
            - USER_EDIT
      tags:
        - "PostStatistic"
      operationId: setVoteStatistic
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Post id
        - name: vote
          in: path
          required: true
          schema:
            type: boolean
            example: true
            description: Vote
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostStatisticGet'
  /v1/poststatistic/{postId}/favorite/{favorite}:
    put:
      summary: Gives favorite for post
      security:
        - BearerAuth:
            - USER_EDIT
      tags:
        - "PostStatistic"
      operationId: setFavoriteStatistic
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Post id
        - name: favorite
          in: path
          required: true
          schema:
            type: boolean
            example: true
            description: Favorite
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostStatisticGet'
  /v1/post/{id}/postobject/{order}:
    post:
      summary: Creates new postobject for unpublished post
      security:
        - BearerAuth:
            - USER_ADD
      tags:
        - "PostObject"
      operationId: createPostObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Post id
        - name: order
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Weight of post object
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File of post object
                description:
                  type: string
                  description: Description of file
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostObjectFullGet'
...
  /v1/profilepicture:
    post:
      summary: Creates new profile picture
      security:
        - BearerAuth:
            - MODERATOR_ADD
      tags:
        - "ProfilePhoto"
      operationId: addProfilePicture
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File of post object
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProfilePictureGet'
  /v1/tag/all:
    post:
      summary: Returns paginated list of tags
      tags:
        - "Tag"
      operationId: getAllTags
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TagPageGet'
...
  /v1/tag/{id}:
    get:
      summary: Returns tag by given id
      tags:
        - "Tag"
      operationId: getTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TagGet'
    delete:
      summary: Deletes tag by given id
      security:
        - BearerAuth:
            - MODERATOR_DELETE
      tags:
        - "Tag"
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Tag id
            example: 1
      responses:
        '200':
          description: OK
...
  /v1/post/{id}/poststatistic/:
    get:
      summary: Returns post statistics
      tags:
        - "PostStatistic"
      operationId: getPostStatistic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Post id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostStatisticGet'
...
components:
  schemas:
...
    PostStatisticGet:
      type: object
      properties:
        postId:
          type: integer
          format: int64
          description: Post statistic id
          example: 1
        seenCount:
          type: integer
          format: int64
          description: Ammount of times this post was seen by unique user
          example: 1
        upVoteCount:
          type: integer
          format: int64
          description: Ammount of positive votes this post have
          example: 2
        downVoteCount:
          type: integer
          format: int64
          description: Ammount of negative votes this post have
          example: 2
        favoriteCount:
          type: integer
          format: int64
          description: Ammount of favorites this post have
          example: 2
    TagGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of tag
          example: 1
        tag:
          type: string
          description: Value of tag
          example: Cat
...
    ProfilePictureGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of profile picture
          example: 1
        content:
          type: string
          format: byte
          description: Base64 string
        mimeType:
          type: string
          description: Mime type of profile picture
          example: image/jpeg
...
    TagFilter:
      type: object
      properties:
        tag:
          type: string
          description: Tag to search for
          example: Cat
          nullable: true
    TagPagingAndSortingRequest:
      type: object
      required:
        - page
        - size
      properties:
        page:
          type: integer
          format: int32
          description: Page to display
          example: 1
        size:
          type: integer
          format: int32
          description: Size of page to display
          example: 1
        sortBy:
          type: string
          description: Sort by field
          example: id
        sortDirection:
          type: string
          description: Sort direction
          example: asc
    TagRequest:
      type: object
      properties:
        pagingAndSorting:
          $ref: '#/components/schemas/TagPagingAndSortingRequest'
        filters:
          $ref: '#/components/schemas/TagFilter'
    TagPageGet:
      type: object
      properties:
        tags:
          type: array
          description: Array containing tags
          items:
            $ref: '#/components/schemas/TagGet'
        totalNumberOfPages:
          type: integer
          format: int32
          description: Total number of pages with current filter applied
          example: 1
        totalNumberOfElements:
          type: integer
          format: int64
          description: Total number of elements that can be queried
          example: 1
        filters:
          $ref: '#/components/schemas/TagFilter'
...
    PostObjectFullGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Id of post object
        postId:
          type: integer
          format: int64
          example: 1
          description: Id of post
        fileName:
          type: string
          example: meme1
          description: Name of the file
        content:
          type: string
          format: byte
          description: Base64 string
        mimeType:
          type: string
          example: image/png
          description: Mime type of file
        description:
          type: string
          example: Some file description
          description: Description of file